\name{wlsObject}
\docType{data}
\alias{wlsObject}
\title{Example object of class wls}
\description{
wlsObject: an example object of class wls. See more: \code{reGenotyper}.
}
\usage{
data(wlsObject)
}

\author{ 
 Yang Li <yang.li@rug.nl> 
}

\examples{
  ##load the data
  data(wlsObject)
  wlsObject
  plot(wlsObject)
}
\keyword{datasets}
